mali_kbase: Improve DVFS utilization accuracy
Get the current time at the same time the utilization registers are
accessed, and disable interrupts on RT while it happens so that nothing can
interrupt the calculation and hurt the accuracy of the numbers.

This ensures that inaccurate DVFS utilization numbers aren't fed into the
governor, which should therefore improve the frequency selections made by
the governor.

Change-Id: Ia64eea829e3fe6acb55457758bc79c3246680b0e
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: minaripenguin <minaripenguin@users.noreply.github.com>